requirements.txt
paho-mqtt==1.6.1
json-logging-py==0.2
python-dateutil==2.8.2
pytest==7.4.3
pytest-cov==4.1.0